Revert "new route"
This reverts commit 0b18d23991e1ed111e7505c98b9d2c54c66960e6.
