+++ src/routes/bocantino/lotes/+page.server.ts
@@ -1,14 +1,13 @@
 import { ingredient_production_service } from '$logic/ingredient-production-service';
 import { product_service } from '$logic/product-service';
-import type { PageServerLoad } from './$types';
-import { redirect } from '@sveltejs/kit';
+import type { PageServerLoad } from '../$types';
 
 export const load: PageServerLoad = async ({ url }) => {
-	throw redirect(300, '/bocantino/lotes/ingresados');
-	//const page = Number(url.searchParams.get('page')) || 0; // cant use ??  because of NaN
-	//const batches = await ingredient_production_service.getBatchesAvailable({ page });
-	//const count_batches = await ingredient_production_service.getCountOfAvailableBatches();
-	//const page_size = ingredient_production_service.PAGE_SIZE;
+	const page = Number(url.searchParams.get('page')) || 0; // cant use ??  because of NaN
+	const batches = await ingredient_production_service.getBatchesAvailable({ page });
+	const count_batches = await ingredient_production_service.getCountOfAvailableBatches();
+	const page_size = ingredient_production_service.PAGE_SIZE;
 
-	//return { batches, count_batches, page_size };
+	return { batches, count_batches, page_size };
 };
+
