+++ src/routes/bocantino/lotes/[id]/+page.server.ts
