+++ src/routes/bocantino/lotes/[id]/+page.svelte
